requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
executing==2.1.0
filelock==3.13.1
Flask==3.1.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
openai==1.65.3
packaging==24.1
parso==0.8.4
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.2.0
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
